pyproject.toml
[tool.poetry]
name = "visual_testing"
version = "0.1.0"
description = "Visual testing POC to demonstrate multiple apparchaes for Visual testing of UI, images and PDF files."
authors = ["Manoj Patil <manoj.patil@atlascopco.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11.0"
selenium = "^4.2.0"
Pillow = "^9.1.1"
numpy = "^1.22.4"
imutils = "^0.5.4"
opencv-python = "^4.5.5"
eyes-selenium = "^5.6.1"
behave = "^1.2.6"
polling = "^0.3.2"
requests = "^2.27.1"
allure-behave = "^2.9.45"
json2html = "^1.3.0"
python-dotenv = "^0.20.0"
pdf2image = "^1.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
